Remove graphs app
